requirements.txt
beautifulsoup4==4.11.1
selenium==3.141.0
pandas==1.1.5
2captcha-python==1.1.2
python-dotenv==0.20.0
Pillow==8.4.0